% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{basic_factor_matrix}
\alias{basic_factor_matrix}
\title{basic_factor_matrix}
\usage{
basic_factor_matrix(k)
}
\arguments{
\item{k}{Number of basic factors.}
}
\value{
A \eqn{2^{k} \times k} matrix containing the basic factors.
}
\description{
Generate the matrix containing `k` basic factors.
}
