% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clear_2fi}
\alias{clear_2fi}
\title{Number of clear 2FI}
\usage{
clear_2fi(X)
}
\arguments{
\item{X}{A numeric matrix representing the design matrix.}
}
\value{
An integer representing the number of clear 2FI in the design.
}
\description{
Compute the number of clear two-factor interactions (2FI) of a
design. A 2FI that is not aliased with any main effect or other 2FI's is
called clear.
}
