% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{custom_design}
\alias{custom_design}
\title{Generate a custom design}
\usage{
custom_design(run_size, added_cols, add_bf = T)
}
\arguments{
\item{run_size}{Integer representing the number of runs in the design.}

\item{added_cols}{Numeric vector holding the numbers of the columns of the
design. It should only contain the added columns, as all the basic factors
are added to the set of columns before generating the design.}

\item{add_bf}{Logical. Add the numbers corresponding to the basic factors to
the column numbers already provided. Default is TRUE.}
}
\value{
A numeric matrix representing the design matrix.
}
\description{
Given a run size and a set of column numbers, generate the
corresponding design matrix.
}
\examples{
custom_design(32, c(7,15,31))
custom_design(16, c(1,2,8,11), add_bf = F)
}
