% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{design}
\alias{design}
\title{design}
\usage{
design(run_size, index)
}
\arguments{
\item{run_size}{An integer specifying the run size of the design. It must be
a power of 2.}

\item{index}{A string representing the index of the design, from the paper.
All strings are of the from "n-p.i" where `n` is the number of factors in the
design, `p` is the number of added factors in the design, and `i` is the rank
of the design in the table.}
}
\value{
A numeric matrix corresponding to the design matrix.
}
\description{
`design` generates the design matrix of a design from the
catalog of Chen, Sun and Wu (1993) or Xu (2009) using the run size and the
index of the design as identifiers.

    The index of a design can be found in the design tables of the
corresponding paper.
}
