% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{design_model_matrix}
\alias{design_model_matrix}
\title{Generate the model matrix for the design with `k` factors and the columns
specified in the vector ´cols´.}
\usage{
design_model_matrix(k, cols)
}
\arguments{
\item{k}{Number of basic factors.}

\item{cols}{Columns used in the design.}
}
\value{
A numeric matrix corresponding to the model matrix of the design.
}
\description{
Generate the model matrix for the design with `k` factors and the columns
specified in the vector ´cols´.
}
