% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance_distribution}
\alias{distance_distribution}
\title{Distance distribution of a matrix}
\usage{
distance_distribution(x)
}
\arguments{
\item{x}{input matrix}
}
\value{
A vector of the distance distribution
}
\description{
Computes the distribution of the hamming distances between
all pairs of rows (even identical) in a matrix. For a matrix with N rows,
there are \eqn{\binom{N}{2}+N} such combinations.
}
\details{
For a matrix \eqn{\mathbf{X}} of size \eqn{(n \times m)}, the distance
distribution is a row-vector \eqn{B} of length `n`, where \eqn{B_i} is
the number of pairs of rows with hamming distance \eqn{i} divided by `n`.
Since pairs of identical rows are considered, \eqn{B_0} will always be `n`.
}
