% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hamming_distance}
\alias{hamming_distance}
\title{Hamming distance between two vectors}
\usage{
hamming_distance(a, b)
}
\arguments{
\item{a}{first vector}

\item{b}{second vector}
}
\value{
An integer
}
\description{
Computes the hamming distance between two vectors \code{a} and
\code{b}. The hamming distance is the number of places where they differ.
Therefore, for two vectors of size n, the distance is an integer between 0
and n.
}
