% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{krawtchouk}
\alias{krawtchouk}
\title{Krawtchouk polynomials}
\usage{
krawtchouk(j, x, n)
}
\arguments{
\item{j}{An integer}

\item{x}{An integer}

\item{n}{An integer}
}
\value{
An integer giving the value of the polynomial evaluated at j, x and
n.
}
\description{
Compute the Krawtchouk polynomial P_j(x,n,s) for s=2, since it
is for two-level designs.
}
