% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mac_williams_transform}
\alias{mac_williams_transform}
\title{MacWilliams Transform}
\usage{
mac_williams_transform(B, N)
}
\arguments{
\item{B}{A numeric vector containing the distance distribution of a design.}

\item{N}{An integer representing the run size of the design.}
}
\value{
A numeric vector of the same size as B.
}
\description{
Compute the MacWiliams transform of the distance distribution
of a design.
}
