% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nchoosek}
\alias{nchoosek}
\title{Binomial coefficient}
\usage{
nchoosek(n, k)
}
\arguments{
\item{n}{An integer, the size of the set of all elements.}

\item{k}{An integer, the number of elements to choose.}
}
\value{
An integer
}
\description{
Compute the number of ways of choosing `k` elements out of a
set of size `n`. Chosen behavior for negative `k` is to return 0. Chosen
behavior for `k` equals to zero is to return 1.
}
