% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num2gen.R
\name{num2gen}
\alias{num2gen}
\title{num2gen}
\usage{
num2gen(n)
}
\arguments{
\item{n}{A numeric or numeric vector.}
}
\value{
A string or character vector holding the generator corresponding to
`n`
}
\description{
`num2gen` converts a number into its corresponding generator.
}
\examples{
num2gen(7)
generators <- num2gen(n = c(13,31,28))
}
