% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wlp}
\alias{wlp}
\title{Word length pattern}
\usage{
wlp(D)
}
\arguments{
\item{D}{A numeric matrix corresponding to the design matrix.}
}
\value{
A numeric vector with size equal to the number of columns of the
original design D, containing the word length pattern.
}
\description{
Compute the Word Length Pattern (WLP) of a design. For this
vector B, B_i represents the number of words of length i.
}
